+++ kubejs/startup_scripts/firmalife/constants.js
@@ -33,7 +33,9 @@ global.FIRMALIFE_DISABLED_ITEMS = [
     'firmalife:irrigation_tank',
 
     // disabled so people use tongs instead
-    'firmalife:peel'
+    'firmalife:peel',
+
+    'firmalife:bucket/chocolate'
 ];
 
 global.FIRMALIFE_HIDED_ITEMS = [